requirements.txt
streamlit
pandas
docxtpl
streamlit-drawable-canvas
Pillow
google-api-python-client
google-auth